refactoring and update-delete post features

+++ backend/routes/users.js
@@ -3,15 +3,29 @@ const User = require("../models/User")
 const Post = require("../models/Post")
 const bcrypt = require("bcrypt")
 
-// UPDATE USER
+
+// Get user
+router.get("/:id", async (req, res) => {
+    try {
+        const user = await User.findById(req.params.id);
+        const { password, ...others } = user._doc
+        res.status(200).json(others)
+
+    } catch(err) {
+        res.status(500).json(err)
+    }
+})
+
+// Updating user
 router.put("/:id", async (req, res) => {
+    
     if(req.body.userId === req.params.id) {
 
-        // if there's a password as well
+        // if we are sending a password as well
         if(req.body.password){
             const salt = await bcrypt.genSalt(10)
-            const newPassword = await bcrypt.hash(req.body.password, salt);
-             req.body.password = newPassword;
+            const updatedPassword = await bcrypt.hash(req.body.password, salt);
+            req.body.password = updatedPassword;
         }
         try {
             const updatedUser = await User.findByIdAndUpdate(req.params.id, {
@@ -25,43 +39,32 @@ router.put("/:id", async (req, res) => {
         }
     }
     else {
-        res.status(401).json("You can update your account only");
+        res.status(401).json("You can update your account only!!");
     }
 })
 
-// DELETE USER
+// Deleting User
 router.delete("/:id", async (req, res) => {
     if(req.body.userId === req.params.id){
         try {
             const user = await User.findById(req.params.id)
-            try {
-                // deleting all its posts
-                await Post.deleteMany({username: user.username})
-                await User.findByIdAndDelete(req.params.id);
-
-                res.status(200).json("User has been deleted!")
-            } catch(err) {
-                res.status(500).json(err);
+            if(user){
+                try {
+                    // deleting all its posts
+                    await Post.deleteMany({username: user.username})
+                    await User.findByIdAndDelete(req.params.id);
+    
+                    res.status(200).json("User has been deleted!")
+                } catch(err) {
+                    res.status(500).json(err);
+                }
             }
-
         } catch(err) {
             res.status(404).json("User not found!")
         }
     }
     else {
         res.status(500).json("You can delete your account only")
-    }
-})
-
-// GET USER
-router.get("/:id", async (req, res) => {
-    try {
-        const user = await User.findById(req.params.id);
-        const { password, ...others } = user._doc
-        res.status(200).json(others)
-
-    } catch(err) {
-        res.status(500).json(err)
     }
 })
